requirements.txt
flask
firebase-admin
python-dotenv
numpy
pandas
scikit-learn
joblib
